+++ bootcfg/http/metadata_test.go
@@ -12,24 +12,32 @@ import (
 	"golang.org/x/net/context"
 
 	"github.com/coreos/coreos-baremetal/bootcfg/storage/storagepb"
-	fake "github.com/coreos/coreos-baremetal/bootcfg/storage/testfakes"
 )
 
 func TestMetadataHandler(t *testing.T) {
+	group := &storagepb.Group{
+		Id:       "test-group",
+		Selector: map[string]string{"mac": "52:54:00:a1:9c:ae"},
+		Metadata: []byte(`{"meta":"data", "etcd":{"name":"node1"},"some":{"nested":{"data":"some-value"}}}`),
+	}
 	logger, _ := logtest.NewNullLogger()
 	srv := NewServer(&Config{Logger: logger})
 	h := srv.metadataHandler()
-	ctx := withGroup(context.Background(), fake.Group)
+	ctx := withGroup(context.Background(), group)
 	w := httptest.NewRecorder()
-	req, _ := http.NewRequest("GET", "/?uuid=a1b2c3d4", nil)
+	req, _ := http.NewRequest("GET", "/?mac=52-54-00-a1-9c-ae", nil)
 	h.ServeHTTP(ctx, w, req)
 	// assert that:
 	// - the Group's custom metadata and selectors are served
 	// - key names are upper case
 	expectedData := map[string]string{
-		"POD_NETWORK":  "10.2.0.0/16",
-		"SERVICE_NAME": "etcd2",
-		"UUID":         "a1b2c3d4",
+		// group metadata
+		"META": "data",
+		"ETCD": "map[name:node1]",
+		"SOME": "map[nested:map[data:some-value]]",
+		// group selector
+		"MAC": "52:54:00:a1:9c:ae",
+		// HACK(dghubble): Not testing query params until #84
 	}
 	assert.Equal(t, http.StatusOK, w.Code)
 	// convert response (random order) to map (tests compare in order)
@@ -57,11 +65,11 @@ func TestMetadataHandler_MetadataEdgeCases(t *testing.T) {
 		w := httptest.NewRecorder()
 		req, _ := http.NewRequest("GET", "/", nil)
 		h.ServeHTTP(ctx, w, req)
-		// assert that:
-		// - the Group's custom metadata is served
+		// assert that each Group's metadata is formatted:
 		// - key names are upper case
+		// - key/value pairs are newline separated
 		assert.Equal(t, http.StatusOK, w.Code)
-		assert.Equal(t, c.expected, w.Body.String())
+		assert.Contains(t, w.Body.String(), c.expected)
 		assert.Equal(t, plainContentType, w.HeaderMap.Get(contentType))
 	}
 }
@@ -85,6 +93,10 @@ func metadataToMap(metadata string) map[string]string {
 		token := scanner.Text()
 		pair := strings.SplitN(token, "=", 2)
 		if len(pair) != 2 {
+			continue
+		}
+		// HACK(dghubble) - Skip map unwinding until #84
+		if pair[0] == "REQUEST" {
 			continue
 		}
 		data[pair[0]] = pair[1]